+++ modules/planning/math/smoothing_spline/spline_2d_kernel.cc
@@ -30,7 +30,7 @@ namespace planning {
 using apollo::common::math::Vec2d;
 
 Spline2dKernel::Spline2dKernel(const std::vector<double>& t_knots,
-                               const std::size_t spline_order)
+                               const std::uint32_t spline_order)
     : t_knots_(t_knots), spline_order_(spline_order) {
   total_params_ =
       (t_knots_.size() > 1 ? 2 * (t_knots_.size() - 1) * spline_order_ : 0);
@@ -78,7 +78,7 @@ const Eigen::MatrixXd& Spline2dKernel::offset() const { return offset_; }
 
 // build-in kernel methods
 void Spline2dKernel::add_derivative_kernel_matrix(const double weight) {
-  for (std::size_t i = 0; i + 1 < t_knots_.size(); ++i) {
+  for (std::uint32_t i = 0; i + 1 < t_knots_.size(); ++i) {
     Eigen::MatrixXd cur_kernel =
         SplineSegKernel::instance()->kernel_derivative(
             spline_order_, t_knots_[i + 1] - t_knots_[i]) *
@@ -92,7 +92,7 @@ void Spline2dKernel::add_derivative_kernel_matrix(const double weight) {
 }
 
 void Spline2dKernel::add_second_order_derivative_matrix(const double weight) {
-  for (std::size_t i = 0; i + 1 < t_knots_.size(); ++i) {
+  for (std::uint32_t i = 0; i + 1 < t_knots_.size(); ++i) {
     Eigen::MatrixXd cur_kernel =
         SplineSegKernel::instance()->kernel_second_order_derivative(
             spline_order_, t_knots_[i + 1] - t_knots_[i]) *
@@ -106,7 +106,7 @@ void Spline2dKernel::add_second_order_derivative_matrix(const double weight) {
 }
 
 void Spline2dKernel::add_third_order_derivative_matrix(const double weight) {
-  for (std::size_t i = 0; i + 1 < t_knots_.size(); ++i) {
+  for (std::uint32_t i = 0; i + 1 < t_knots_.size(); ++i) {
     Eigen::MatrixXd cur_kernel =
         SplineSegKernel::instance()->kernel_third_order_derivative(
             spline_order_, t_knots_[i + 1] - t_knots_[i]) *
@@ -128,13 +128,13 @@ bool Spline2dKernel::add_reference_line_kernel_matrix(
     return false;
   }
 
-  for (std::size_t i = 0; i < t_coord.size(); ++i) {
+  for (std::uint32_t i = 0; i < t_coord.size(); ++i) {
     double cur_index = find_index(t_coord[i]);
     double cur_rel_t = t_coord[i] - t_knots_[cur_index];
     // update offset
     double offset_coef_x = -ref_points[i].x() * weight;
     double offset_coef_y = -ref_points[i].y() * weight;
-    for (std::size_t j = 0; j < spline_order_; ++j) {
+    for (std::uint32_t j = 0; j < spline_order_; ++j) {
       offset_(j + (2 * cur_index) * spline_order_, 0) = offset_coef_x;
       offset_(j + (2 * cur_index + 1) * spline_order_, 0) = offset_coef_y;
       offset_coef_x *= cur_rel_t;
@@ -146,13 +146,13 @@ bool Spline2dKernel::add_reference_line_kernel_matrix(
 
     double cur_t = 1.0;
     std::vector<double> power_t;
-    for (std::size_t n = 0; n + 1 < 2 * spline_order_; ++n) {
+    for (std::uint32_t n = 0; n + 1 < 2 * spline_order_; ++n) {
       power_t.emplace_back(cur_t);
       cur_t *= cur_rel_t;
     }
 
-    for (std::size_t r = 0; r < spline_order_; ++r) {
-      for (std::size_t c = 0; c < spline_order_; ++c) {
+    for (std::uint32_t r = 0; r < spline_order_; ++r) {
+      for (std::uint32_t c = 0; c < spline_order_; ++c) {
         ref_kernel(r, c) = power_t[r + c];
       }
     }
@@ -166,10 +166,10 @@ bool Spline2dKernel::add_reference_line_kernel_matrix(
   return true;
 }
 
-std::size_t Spline2dKernel::find_index(const double t) const {
+std::uint32_t Spline2dKernel::find_index(const double t) const {
   auto upper_bound = std::upper_bound(t_knots_.begin() + 1, t_knots_.end(), t);
-  return std::min(t_knots_.size() - 1,
-                  static_cast<std::size_t>(upper_bound - t_knots_.begin())) -
+  return std::min(static_cast<std::uint32_t>(t_knots_.size() - 1),
+                  static_cast<std::uint32_t>(upper_bound - t_knots_.begin())) -
          1;
 }
