+++ modules/planning/math/smoothing_spline/spline_2d_solver.cc
@@ -25,7 +25,7 @@ namespace apollo {
 namespace planning {
 
 Spline2dSolver::Spline2dSolver(const std::vector<double>& t_knots,
-                               const std::size_t order)
+                               const std::uint32_t order)
     : spline_(t_knots, order),
       kernel_(t_knots, order),
       constraint_(t_knots, order) {}
@@ -60,7 +60,7 @@ bool Spline2dSolver::solve() {
     return false;
   }
 
-  const std::size_t spline_order = spline_.spline_order();
+  const std::uint32_t spline_order = spline_.spline_order();
   const Eigen::MatrixXd solved_params = qp_solver_->params();
   return spline_.set_splines(solved_params, spline_order);
 }
